+++ .vulture_whitelist.py
@@ -33,3 +33,4 @@ _attr_state_class  # unused attribute
 _attr_options  # unused attribute
 _attr_translation_key  # unused attribute
 use_as_default  # unused arg required by pymodbus' ModbusSlaveContext.setValues signature
+async_select_option  # unused method required by Home Assistant's SelectEntity